requirements.txt
numpy
scipy
matplotlib
seaborn
scikit-learn
faiss-cpu
